requirements.txt
SpeechRecognition>=3.10.0
pyttsx3
pyaudio
wikipedia
requests
google-generativeai
pywhatkit
AppOpener
python-dotenv
setuptools
flask
flask-socketio
psutil
gTTS
pygame
pyautogui
edge-tts
googlesearch-python
beautifulsoup4
asyncio
mss
Pillow
spotipy
